# docs/openapi.yaml
openapi: 3.1.0
info:
  title: http402-forge-api
  version: 0.1.0
  description: |
    Forge marketplace — list, preview, pay-to-download via x402 v2.
    JSON responses use **camelCase** field names.
servers:
  - url: https://api.http402.trade
  - url: http://127.0.0.1:8092
components:
  schemas:
...
    PaymentRequired:
      type: object
      description: x402 v2 payment required body (402 response)
      properties:
        x402Version: { type: integer }
        resource: { type: object }
        accepts:
          type: array
          items: { type: object }
        extensions: { type: object }
    ErrorBody:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
paths:
...
  /api/v1/listings/{id}/download:
    get:
      summary: Pay-to-download (402 then streamed asset)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: PAYMENT-SIGNATURE
          schema: { type: string }
          description: Base64 JSON proof from facilitator build+sign flow
      responses:
        "402":
          description: Payment required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentRequired" }
        "200":
          description: Asset file stream
        "403":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
...